backend/requirements.txt
# Trading AI SHER - Python Backend
# FastAPI + ML + Trading Engine

# ===== Core Framework =====
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# ===== Database =====
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0
asyncpg==0.29.0
greenlet==3.0.3

# ===== Authentication =====
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0

# ===== HTTP & WebSocket =====
httpx==0.26.0
websockets==12.0
aiohttp==3.9.1

# ===== Machine Learning =====
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.3.0
torch==2.2.0
ta==0.11.0

# ===== Data Processing =====
python-dotenv==1.0.0
pyyaml==6.0.1
orjson==3.9.12

# ===== Broker Integration =====
smartapi-python==1.0.0

# ===== Utilities =====
python-dateutil==2.8.2
pytz==2024.1
loguru==0.7.2
rich==13.7.0

# ===== Testing =====
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# ===== Development =====
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# ===== Production =====
gunicorn==21.2.0
redis==5.0.1
